+++ Tezo.Todo.Repository/Tezo.Todo.Repository.csproj
@@ -1,4 +1,4 @@
-﻿<Project Sdk="Microsoft.NET.Sdk.Web">
+<Project Sdk="Microsoft.NET.Sdk.Web">
 
   <PropertyGroup>
     <TargetFramework>net8.0</TargetFramework>
@@ -8,14 +8,12 @@
   </PropertyGroup>
 
   <ItemGroup>
-    <PackageReference Include="AutoMapper.Extensions.Microsoft.DependencyInjection" Version="12.0.1" />
     <PackageReference Include="Swashbuckle.AspNetCore" Version="6.4.0" />
   </ItemGroup>
 
   <ItemGroup>
     <ProjectReference Include="..\Tezo.Todo.Data\Tezo.Todo.Data.csproj" />
     <ProjectReference Include="..\Tezo.Todo.Models\Tezo.Todo.Models.csproj" />
-    <ProjectReference Include="..\Tezo.Todo.Repository\Tezo.Todo.Repository.csproj" />
   </ItemGroup>
 
 </Project>
