+++ Tezo.Todo.Services/Services/UserService.cs
@@ -5,59 +5,39 @@ using Tezo_Todo.Services.Interfaces;
 namespace Tezo_Todo.Services
 {
     public class UserService : IUserServices
-    {
-        private readonly TodoAPIDbContext _dbContext;
+    {     
+        private readonly UserRepository userRepository;
 
         public UserService(TodoAPIDbContext dbContext)
         {
-            this._dbContext = dbContext;
+            userRepository = new UserRepository(dbContext);
         }
 
         public List<User> GetAllUser()
         {
-            return _dbContext.User.ToList();
+            return userRepository.GetAllUser();
         }
 
         public User AddUser(User user)
         {
-            _dbContext.User.AddAsync(user);
-            _dbContext.SaveChangesAsync();
-
-            return user;
+            return userRepository.AddUser(user);
         }
 
-        public async Task<User> GetUserById(Guid id)
+        public Task<User> GetUserById(Guid id)
         {
-            var user = await _dbContext.User.FindAsync(id);
-            return user;
+            return userRepository.GetUserById(id);
         }
 
 
-        public async Task<bool> UpdateUser(Guid id, User user)
+        public  Task<bool> UpdateUser(Guid id, User user)
         {
-            try
-            {
-                var userData = _dbContext.User.Update(user);
-                _dbContext.SaveChanges();
-                return true;
-            }
-            catch (Exception ex)
-            {
-                return false;
-            }
-
+            return userRepository.UpdateUser(id, user);
         }
 
 
-        public async Task<User> DeleteUser(Guid id)
+        public  Task<User> DeleteUser(Guid id)
         {
-            var user = await GetUserById(id);
-            if (user != null)
-            {
-                _dbContext.User.Remove(user);
-                _dbContext.SaveChangesAsync();
-            }
-            return user;
+            return userRepository.DeleteUser(id);
         }
