--- Tezo_Todo/Models/User.cs
@@ -1,29 +1,0 @@
-﻿//using System.ComponentModel.DataAnnotations;
-
-//namespace Tezo_Todo.Models
-//{
-//    public class User
-//    {
-
-//        [Key]
-//        public Guid Id { get; set; }
-//        public string UniqueName { get; set; }
-//        public string FirstName { get; set; }
-//        public string LastName { get; set; }
-//        public string Password { get; set; }
-
-//        public User()
-//        {
-
-//        }
-
-//        public User(Guid id, string uniqueName, string firstName, string lastName, string password)
-//        {
-//            Id = id;
-//            UniqueName = uniqueName;
-//            FirstName = firstName;
-//            LastName = lastName;
-//            Password = password;
-//        }
-//    }
-//}
